+++ SpringBoot_JEE_Projet/src/main/java/org/thes/entities/LigneCommande.java
@@ -21,7 +21,7 @@ import lombok.ToString;
 
 public class LigneCommande {
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id_ligneCommande;
+	private Long id_ligneCommande;
 	private int quantite;
 	private double prixLigne;
